db/requirements.txt
pymongo
pydanticcoverage==7.8.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.5.0
pytest==8.3.5
email-validator